Changed the if statement in the app component root html file so that it uses the activeGym instead of the gymSelected property.
Also removed the gym selectedProperty from the gym service to make thing simpler and as it is now redundant.

Moved all the database setting up methods to a text file so that the service is smaller.

Additionally extended the gym object so it can now parse population data taken from the service. The gym service can now fetch the appropriate login data for a specific gym.

Created a method that draws a graph with data from this but it is not working because the reference to active gym is not working!